Refresh token

+++ src/main/java/com/vti/testing/service/UserService.java
@@ -1,53 +1,93 @@
 package com.vti.testing.service;
-
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import com.vti.testing.dto.UserDTO;
+import com.vti.testing.entity.Role;
 import com.vti.testing.entity.User;
+import com.vti.testing.exception.custom_exception.AlreadyExistEx;
 import com.vti.testing.exception.custom_exception.NotFoundEx;
+import com.vti.testing.formcreate.FormUserCreate;
+import com.vti.testing.repository.IRoleRepository;
 import com.vti.testing.repository.IUserRepository;
-
+import com.vti.testing.responseobj.ResponseObj;
+import com.vti.testing.service.interfaces.IUserService;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 @Service
 public class UserService implements IUserService {
-
+	private final String PASSWORD_PATTERN =
+			"^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#&()–[{}]:;',?/*~$^+=<>]).{8,20}$";
+	private final Pattern pattern = Pattern.compile(PASSWORD_PATTERN);
+	public boolean validatePassWord(String password) {
+		Matcher matcher = pattern.matcher(password);
+		return matcher.matches();
+	}
 	@Autowired
 	private IUserRepository userRepository;
 	@Autowired
+	private IRoleRepository roleRepository;
+	@Autowired
 	private ModelMapper modelMapper;
-	
+	@Autowired
+	private BCryptPasswordEncoder passwordEncoder;
 	@Override
 	public User getByUserName(String userName) {
-		return userRepository.findByUserName(userName);
+		if(userRepository.existsByUserName(userName)){
+			return userRepository.findByUserName(userName);
+		}
+		throw new NotFoundEx("This user: " + userName + " is not found!");
 	}
 
 	@Override
 	public UserDTO getUserById(int id) {
 		Optional<User> user = userRepository.findById(id);
 		if (user.isPresent()) {
-			UserDTO dto = modelMapper.map(user.get(), UserDTO.class);
-			return dto;
+			return modelMapper.map(user.get(), UserDTO.class); // dto object
 		}
-		throw new NotFoundEx("Cannot find user with this id " + id );
+		throw new NotFoundEx("Cannot find user with id = " + id );
 	}
 
-//	@Override
-//	public void newUser(FormUserCreate newUser) {
-//		User user = new User(newUser.getUserName(), newUser.getPassWord(), newUser.getActive(), newUser.getRole());
-//		userRepository.save(user);
-//
-//	}
+	@Override
+	public ResponseObj newUser(FormUserCreate newUser) throws Exception {
+		String userName = newUser.getUserName();
+		String passWord = newUser.getPassWord();
+		if(userRepository.existsByUserName(userName)){
+			throw new AlreadyExistEx("This user: " + userName + " has been used!");
+		}
+		if (!validatePassWord(passWord)){
+			throw new Exception("Password is not valid, please check your password");
+		}
+		// Encode passWord
+		String passEncode = passwordEncoder.encode(passWord);
+		newUser.setPassWord(passEncode);
+		User user = new User(userName, newUser.getPassWord());
+		Role role = roleRepository.findByRoleName("ROLE_USER");
+		List<Role> newUserRoles = new ArrayList<Role>();
+			newUserRoles.add(role); // role default = ROLE_USER
+		user.setRoles(newUserRoles);
+		user.setActive("1");
+		userRepository.save(user);
+		return new ResponseObj("200", "New User has been created", userName);
+	}
 
 	@Override
 	public List<UserDTO> getAllUsers() {
-		List<User> entitys = userRepository.findAll();
-		List<UserDTO> dtos = modelMapper.map(entitys,
+		List<User> entities = userRepository.findAll();
+		return modelMapper.map(entities,
+				new TypeToken<List<UserDTO>>(){}.getType()); // return list dtos
+	}
+
+	@Override
+	public List<UserDTO> getAllActiveUser() {
+		List<User> entities = userRepository.getAllActiveUser();
+		return modelMapper.map(entities,
 				new TypeToken<List<UserDTO>>(){}.getType());
-		return dtos;
 	}
 
 	@Override
@@ -59,8 +99,22 @@ public class UserService implements IUserService {
 	}
 
 	@Override
-	public void deleteUser(int id) {
+	public ResponseObj deleteUser(int id) {
+		if (userRepository.existsById(id)){
+			String userNameDelete = userRepository.findById(id).get().getUserName();
 			userRepository.deleteById(id);
+			return new ResponseObj("200", "Delete Success !", userNameDelete);
+		}
+		throw new NotFoundEx("This user is not exists with to delete");
+
+	}
+
+	@Override
+	public UserDTO getById(int id) {
+		if(!userRepository.existsById(id)){
+			throw new AlreadyExistEx("This user is not exists");
+		}
+		return modelMapper.map(userRepository.findById(id), UserDTO.class);
 	}
 
 }
